% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_logo.R
\name{add_dn_logo}
\alias{add_dn_logo}
\title{Add a logo to the current plot}
\usage{
add_dn_logo(g)
}
\arguments{
\item{g}{ggplot obj}
}
\value{
Add the logo in the below right side
}
\description{
Add a logo to the current plot
}
