% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{dn_color}
\alias{dn_color}
\title{Get the hex code for a dn color}
\usage{
dn_color(color_name)
}
\arguments{
\item{color_name}{The name of the color.}
}
\value{

}
\description{
Get the hex code for a dn color
}
