% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{dn_colors}
\alias{dn_colors}
\title{Data Network colors}
\format{
An object of class \code{character} of length 7.
}
\usage{
dn_colors
}
\description{
Data Network colors
}
\keyword{datasets}
