% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_dn}
\alias{scale_color_dn}
\title{scale_color_dn}
\usage{
scale_color_dn(n, type = "discrete", reverse = FALSE, ...)
}
\arguments{
\item{n}{number of colors}

\item{type}{discrete or continuous}

\item{reverse}{reverse order, Default: FALSE}

\item{...}{further params}
}
\description{
scale_color_dn
}
