% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_dn.R
\name{dnGrob}
\alias{dnGrob}
\title{dn geom}
\usage{
dnGrob(
  x,
  y,
  size,
  theme,
  geom_key = list(dn_logo = "please insert logo", google =
    "http://logok.org/wp-content/uploads/2015/09/Google-logo-2015-G-icon.png")
)
}
\arguments{
\item{x}{raster grob x dim}

\item{y}{raster grob y dim}

\item{size}{logo size}

\item{theme}{inherits theme}

\item{geom_key}{which logo}
}
\description{
most of the code is from emoGG https://github.com/dill/emoGG
}
