% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{get_dn_colors}
\alias{get_dn_colors}
\title{Extract Data Network colors from hex codes}
\usage{
get_dn_colors(...)
}
\arguments{
\item{...}{Names of colors in \code{dn_colors}.}
}
\description{
Extract Data Network colors from hex codes
}
