% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\name{get_dn_hex}
\alias{get_dn_hex}
\title{Get the hex code for a dn color palette}
\usage{
get_dn_hex(color_name)
}
\arguments{
\item{color_name}{the color name in dn_colors}
}
\value{

}
\description{
Get the hex code for a dn color palette
}
