% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputs.R
\name{save_svg}
\alias{save_svg}
\title{Save a plot as a svg}
\usage{
save_svg(plot, filename, width, height)
}
\arguments{
\item{plot}{A plot object.}

\item{filename}{Filename to save to an svg.}

\item{width}{Width (inches) of the image}

\item{height}{Height (inches) of the image.}
}
\description{
Save a plot as a svg
}
