% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_fill_dn}
\alias{scale_fill_dn}
\title{Fill scale for dn colors}
\usage{
scale_fill_dn(palette = "seven", discrete = TRUE, reverse = FALSE, ...)
}
\arguments{
\item{palette}{Name of palette in \code{dn_palettes}.}

\item{discrete}{Boolean to indicate if color aesthetic is discrete.}

\item{reverse}{Boolean to indicate whether palette should be reversed.}

\item{...}{Additional arguments passed to \code{discrete_scale} or
\code{scale_color_gradientn}, depending on the value of \code{discrete}.}
}
\description{
Fill scale for dn colors
}
