% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{set_dn_family}
\alias{set_dn_family}
\title{Set the font family used for all supported text elements}
\usage{
set_dn_family(family, title_family = NULL)
}
\arguments{
\item{family}{The name of the font family to use.}

\item{title_family}{The name of the font family to use for the main chart
title. This is an optional argument that lets you use a different font
family for the main title. The argument is ignored if the value is NULL.
The default value is NULL.}
}
\description{
The font family used for each text element is controlled through a set of
options. This function sets each of these options to the same family,
allowing you to quickly customize the font family used by the theme.
Individual family options can be set with \code{options}.
}
