% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scales.R
\name{scale_color_dn}
\alias{scale_color_dn}
\title{scale_color_dn}
\usage{
scale_color_dn(n, type = "discrete", reverse = FALSE, ...)
}
\description{
scale_color_dn
}
\examples{

library(ggplot2)
ggplot(airquality, aes(x = Day, y = Temp,
     group = as.factor(Month), color = as.factor(Month))) +
     geom_point(size = 2.5) +
     scale_color_dn()
}
