% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_dn.R
\name{dnGrob}
\alias{dnGrob}
\title{dn geom}
\usage{
dnGrob(
  x,
  y,
  size,
  theme,
  geom_key = list(dn_logo = "please insert logo", etsy =
    "https://d3rnbxvnd0hlox.cloudfront.net/images/channels/64/icons/large.png", facebook
    = "http://logok.org/wp-content/uploads/2014/10/Facebook-logo-f.png", twitter =
    "http://logok.org/wp-content/uploads/2014/08/Twitter-logo-bird_logo_2012-1024x768.png",
    google = "http://logok.org/wp-content/uploads/2015/09/Google-logo-2015-G-icon.png",
    X23andme = "https://www.23andme.com/static/img/logos/meta_logo.5f584d06594e.png")
)
}
\description{
most of the code is from emoGG https://github.com/dill/emoGG
}
\author{
Ricardo Bion
}
